+++ src/control_plane/project_infrastructure/mod.rs
@@ -1,9 +1,11 @@
 mod plan_soketi_project_resources;
+mod plan_typesense_project_resources;
 mod prepare_project_services;
 mod prepared_project_service;
 mod project_service_preparation_error;
 
 pub(crate) use plan_soketi_project_resources::plan_soketi_project_resources;
+pub(crate) use plan_typesense_project_resources::plan_typesense_project_resources;
 pub(crate) use prepare_project_services::prepare_project_services;
 pub(crate) use prepared_project_service::PreparedProjectService;
 pub(crate) use project_service_preparation_error::ProjectServicePreparationError;